requirements.txt
Django==2.2.3
pytz==2019.1
sqlparse==0.3.0
psycopg2-binary==2.7.7
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
